#my first app #Search works #bug1. setState is undefined, fixed by add a bind function to this function or in template #according to this: https://codedump.io/share/9D0ewVQwxh1n/1/react-ref-and-setstate-not-working-with-es6 ## this.changeContent needs to be bound to the component instance via this.changeContent.bind(this) before being passed as the onChange prop, otherwise the this variable in the body of the function will not refer to the component instance but to window. See Function::bind.
When using React.createClass instead of ES6 classes, every non-lifecycle method defined on a component is automatically bound to the component instance. See Autobinding.

Be aware that binding a function creates a new function. You can either bind it directly in render, which means a new function will be created every time the component renders, or bind it in your constructor, which will only fire once.
#